+++ src/inmanta/data/__init__.py
@@ -32,11 +32,10 @@ import asyncpg
 from asyncpg.protocol import Record
 
 import inmanta.db.versions
-from inmanta import const, util
+from inmanta import const, resources, util
 from inmanta.const import DONE_STATES, UNDEPLOYABLE_NAMES, AgentStatus, ResourceState
 from inmanta.data import model as m
 from inmanta.data import schema
-from inmanta.resources import Id
 from inmanta.server import config
 from inmanta.types import JsonType, PrimitiveTypes
 
@@ -1913,10 +1912,10 @@ class Resource(BaseDocument):
         query = "SELECT * FROM " + cls.table_name() + " WHERE environment=$1 AND attribute_hash IN " + hashes_as_str
         values = [cls._get_value(environment)] + [cls._get_value(h) for h in hashes]
         result = await cls._fetch_query(query, *values)
-        resources = []
+        resources_list = []
         for res in result:
-            resources.append(cls(from_postgres=True, **res))
-        return resources
+            resources_list.append(cls(from_postgres=True, **res))
+        return resources_list
 
     @classmethod
     async def get_resources(
@@ -2038,7 +2037,7 @@ class Resource(BaseDocument):
             async with con.transaction():
                 async for record in con.cursor(query, *values):
                     resource_id = record["resource_id"]
-                    parsed_id = Id.parse_id(resource_id)
+                    parsed_id = resources.Id.parse_id(resource_id)
                     result.append(
                         {
                             "resource_id": resource_id,
@@ -2061,7 +2060,7 @@ class Resource(BaseDocument):
             (filter_statement, values) = cls._get_composed_filter(environment=environment, model=version)
 
         query = f"SELECT * FROM {Resource.table_name()} WHERE {filter_statement}"
-        resources = []
+        resources_list = []
         async with cls._connection_pool.acquire() as con:
             async with con.transaction():
                 async for record in con.cursor(query, *values):
@@ -2069,12 +2068,12 @@ class Resource(BaseDocument):
                         record = dict(record)
                         record["attributes"] = json.loads(record["attributes"])
                         record["id"] = record["resource_version_id"]
-                        parsed_id = Id.parse_id(record["resource_version_id"])
+                        parsed_id = resources.Id.parse_id(record["resource_version_id"])
                         record["resource_type"] = parsed_id.entity_type
-                        resources.append(record)
+                        resources_list.append(record)
                     else:
-                        resources.append(cls(from_postgres=True, **record))
-        return resources
+                        resources_list.append(cls(from_postgres=True, **record))
+        return resources_list
 
     @classmethod
     async def get_resources_for_version_raw(cls, environment, version, projection):
@@ -2114,7 +2113,7 @@ class Resource(BaseDocument):
 
     @classmethod
     def new(cls, environment: uuid.UUID, resource_version_id: m.ResourceVersionIdStr, **kwargs: Any) -> "Resource":
-        vid = Id.parse_id(resource_version_id)
+        vid = resources.Id.parse_id(resource_version_id)
 
         attr = dict(
             environment=environment,